requirements.txt
# MLOps and ML Libraries
mlflow==2.16.2
scikit-learn
pandas
numpy
keras
tensorflow
hyperopt

# Data Engineering and Workflow Orchestration
apache-airflow
apache-airflow-providers-http

# Additional utilities
requests
jupyter